secend change

+++ BuisnesEntity/Entities/ProductE.cs
@@ -11,7 +11,7 @@ namespace BuisnesEntityLayer.Entities
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
 
-        public float Price { get; set; } 
+        public long Price { get; set; } 
         public int Code { get; set; } 
     }
 }
